dossier3/venise_hotel/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Mail
python-dotenv
PyMySQL
